[FIX] aiplanner 모듈에서 사용하는 Long 타입을 모두 Int 타입으로 변경

+++ feature/aiplanner/src/main/kotlin/com/miruni/feature/aiplanner/domain/model/AiPlan.kt
@@ -1,7 +1,7 @@
 package com.miruni.feature.aiplanner.domain.model
 
 data class AiPlan(
-    val aiPlanId: Long, // 세부 일정 ID,
+    val aiPlanId: Int, // 세부 일정 ID,
     val scheduledDate: String, // 세부 일정 진행 날짜
     val startTime: String, // 시작 예정 시간
     val endTime: String, // 종료 예정 시간
